Enabled Flake8 warnings about imports and fixed them
