+++ cozy/main.py
@@ -10,14 +10,12 @@ import argparse
 import datetime
 import pickle
 
-from cozy import syntax
 from cozy import parse
 from cozy import codegen
 from cozy import common
 from cozy import typecheck
 from cozy import desugar
 from cozy import syntax_tools
-from cozy import handle_tools
 from cozy import invariant_preservation
 from cozy import synthesis
 from cozy import opts
