+++ cozy/syntax_tools.py
@@ -40,24 +40,6 @@ def mk_lambda(t, l):
 def compose(f1 : target_syntax.ELambda, f2 : target_syntax.ELambda) -> target_syntax.ELambda:
     return mk_lambda(f2.arg.type, lambda v: f1.apply_to(f2.apply_to(v)))
 
-_SCALAR_TYPES = set((
-    syntax.TBool,
-    syntax.TString,
-    syntax.TNative,
-    syntax.THandle,
-    syntax.TEnum))
-def is_scalar(t : syntax.Type):
-    from cozy.typecheck import is_numeric
-    if is_numeric(t):
-        return True
-    if type(t) in _SCALAR_TYPES:
-        return True
-    if isinstance(t, syntax.TTuple):
-        return all(is_scalar(tt) for tt in t.ts)
-    if isinstance(t, syntax.TRecord):
-        return all(is_scalar(tt) for (f, tt) in t.fields)
-    return False
-
 class BottomUpExplorer(common.Visitor):
     def visit_ADT(self, x):
         new_children = tuple(self.visit(child) for child in x.children())
