+++ cozy/synthesis/__init__.py
@@ -1,2 +1,7 @@
-from .impls import Implementation, construct_initial_implementation
-from .high_level_interface import improve_implementation
+from . import impls
+from . import high_level_interface
+
+# re-export the most important functions and types
+Implementation                   = impls.Implementation
+construct_initial_implementation = impls.construct_initial_implementation
+improve_implementation           = high_level_interface.improve_implementation
