--- tests/handle_tools.py
@@ -1,10 +1,0 @@
-import unittest
-
-from cozy.target_syntax import *
-from cozy.typecheck import retypecheck
-from cozy.syntax_tools import pprint, all_exps
-from cozy.solver import satisfy
-from cozy.evaluation import eval
-
-class TestHandleTools(unittest.TestCase):
-    pass
